+++ src/provider/Def/gotoIncludeDef.ts
@@ -19,77 +19,79 @@ function gotoIncludeDefShowInfo(): void {
     }
 }
 
-function gotoIncludeDefWithTry(mayPath: string, lStr: string, line: number): vscode.LocationLink | null {
-    const pathListUp: string[] = [...pm.DocMap.keys()];
-    // eslint-disable-next-line no-magic-numbers
-    if (Math.random() > 0.3) {
-        pathListUp.reverse();
-    }
-    for (const pathUp of pathListUp) {
-        if (pathUp.toUpperCase().endsWith(mayPath.toUpperCase())) {
-            const col0: number = lStr.length - lStr
-                .replace(/^\s*#Include(?:Again)?\s+/iu, '')
-                .replace(/\*i\s+/iu, '')
-                .length;
-            return {
-                originSelectionRange: new vscode.Range(
-                    new vscode.Position(line, col0),
-                    new vscode.Position(line, lStr.length),
-                ),
-                targetUri: vscode.Uri.file(pathUp),
-                targetRange: new vscode.Range(
-                    new vscode.Position(0, 0),
-                    new vscode.Position(0, 0),
-                ),
-                /**
-                 * The span of this link.
-                 */
-                // targetSelectionRange?: Range,
-            };
+function gotoIncludeDefWithTry(
+    mayPath: string,
+    lStr: string,
+    line: number,
+): vscode.LocationLink[] | null {
+    const mayPath2 = mayPath.replace(/.*[ \\/]/u, '').toUpperCase();
+    const col0: number = lStr.length - lStr
+        .replace(/^\s*#Include(?:Again)?\s+/iu, '')
+        .replace(/\*i\s+/iu, '')
+        .length;
+    const originSelectionRange = new vscode.Range(
+        new vscode.Position(line, col0),
+        new vscode.Position(line, lStr.length),
+    );
+
+    const need: vscode.LocationLink[] = [];
+    for (const pathUp of pm.DocMap.keys()) {
+        const p0: string = pathUp.replace(/.*[ \\/]/u, '').toUpperCase();
+        if (p0 === mayPath2) {
+            const start = new vscode.Position(0, 0);
+            need.push(
+                {
+                    originSelectionRange,
+                    targetUri: vscode.Uri.file(pathUp),
+                    targetRange: new vscode.Range(start, start),
+                },
+            );
         }
     }
 
-    return null;
+    return need;
 }
 
-export function gotoIncludeDef(AhkFileData: TAhkFileData, position: vscode.Position): vscode.LocationLink | null {
+export function gotoIncludeDef(AhkFileData: TAhkFileData, position: vscode.Position): vscode.LocationLink[] | null {
     const { DocStrMap, AST } = AhkFileData;
     const { line } = position;
     const { detail, lStr } = DocStrMap[line];
 
-    if (detail.includes(EDetail.isDirectivesLine) && (/^\s*#Include(?:Again)?\s/iu).test(lStr)) {
-        for (const ahkInclude of collectInclude(AST)) {
-            if (ahkInclude.range.contains(position)) {
-                const { rawData } = ahkInclude;
-                const { type, mayPath } = rawData;
-                if (AIncludePathKnownList.includes(type)) {
-                    const col0: number = lStr.length - lStr
-                        .replace(/^\s*#Include(?:Again)?\s+/iu, '')
-                        .replace(/\*i\s+/iu, '')
-                        .length;
-                    return {
-                        originSelectionRange: new vscode.Range(
-                            new vscode.Position(line, col0),
-                            new vscode.Position(line, lStr.length),
-                        ),
-                        targetUri: vscode.Uri.file(mayPath),
-                        targetRange: new vscode.Range(
-                            new vscode.Position(0, 0),
-                            new vscode.Position(0, 0),
-                        ),
-                        /**
-                         * The span of this link.
-                         */
-                        // targetSelectionRange?: Range,
-                    };
-                }
-                if (type === EInclude.isUnknown) {
-                    return gotoIncludeDefWithTry(mayPath, lStr, line);
-                }
-                gotoIncludeDefShowInfo();
-                break;
-            }
+    if (!(detail.includes(EDetail.isDirectivesLine) && (/^\s*#Include(?:Again)?\s/iu).test(lStr))) return null;
+
+    for (const ahkInclude of collectInclude(AST)) {
+        if (!ahkInclude.range.contains(position)) continue;
+
+        const { rawData } = ahkInclude;
+        const { type, mayPath } = rawData;
+        if (AIncludePathKnownList.includes(type)) {
+            const col0: number = lStr.length - lStr
+                .replace(/^\s*#Include(?:Again)?\s+/iu, '')
+                .replace(/\*i\s+/iu, '')
+                .length;
+            // trim start
+            const start = new vscode.Position(0, 0);
+
+            return [{
+                originSelectionRange: new vscode.Range(
+                    new vscode.Position(line, col0),
+                    new vscode.Position(line, lStr.length),
+                ),
+                targetUri: vscode.Uri.file(mayPath),
+                targetRange: new vscode.Range(start, start),
+                /**
+                 * The span of this link.
+                 */
+                // targetSelectionRange?: Range,
+            }];
         }
+        if (type === EInclude.isUnknown) {
+            return gotoIncludeDefWithTry(mayPath, lStr, line);
+        }
+        // not find
+        gotoIncludeDefShowInfo();
+        break;
     }
+
     return null;
 }